chore: apply deprecated postfix to files that will be deprecated soon
